Weather now appears in the console, but does not render on the DOM. Code I was working on is now commented out in WeatherPreveiw.js
