+++ scripts/parks/ParkDetails.js
@@ -1,5 +1,7 @@
+//import { WeatherSelect } from "../weather/WeatherSelect.js"
 import {useParks} from "./ParkProvider.js"
 import { parkSelect } from "./ParkSelect.js"
+
 useParks
 
 const eventHub = document.querySelector("#mainContainer")
